Split test_deployment.py
